combined DiscoveryEndpoint and DiscoveryUrl
